requirements.txt
playwright>=1.44.0
pandas>=2.2.0
openpyxl>=3.1.0
pyyaml>=6.0.0
